🌱 Split server_ai.go into domain-specific files
Delete the 2004-line pkg/agent/server_ai.go monolith and replace
with 4 focused domain files:

- server_ai_websocket.go: WebSocket upgrade, message dispatch,
  health/clusters/kubectl message handlers, shared constants
  and types (452 lines)
- server_ai_chat.go: Chat streaming via LLM providers, cancel
  handling (WS + HTTP), non-streaming fallback, agent listing
  and selection, error classification (819 lines)
- server_ai_mixed.go: Mixed-mode chat orchestration (thinking
  agent + execution agent), tool-capability detection, Claude
  availability checks (391 lines)
- server_ai_tokens.go: Session token quota, usage tracking,
  persistence to disk, command extraction from LLM responses
  (384 lines)

All methods remain on *Server in the same package — zero interface
or behavior changes.

Fixes #10415
